Tuning Profile Prototype (#63)
* Allow setting of default value in `Karams().get()`

Change `soft` argument to `default`, which returns `default` if `key`
is not found.

* Add instrument to modify PIF values in each frame

When using LatControlPID, the internal values can be modified by
controlling a new `__k_aggro_factor` multiplier.  The calculation is
done in each frame to allow realtime update of the factor.

* Add kommuState event with aggroLevel field

`aggroLevel` can be updated in realtime by publishing to kommuState.  An
`AGGROMAP` dict will map `aggroLevel` to `aggroFactor`.

* Add simple tune profile UI

Add five buttons that publishes kommuState aggroLevel.

* Persist aggroLevel into Karams

Check if aggroLevel has changed during each frame, update and persist it
if different.

NOTE: Updating Karams from ui should be better but require JSON handling
and possibly need to deal with slow operation too.  Take note of the
frame latency and consider this alternative if needed.