Refactor `controlsd.py` for correct check order
* Refactored code to ensure ALC/blinker/lat_active are checked and used
within the same cycle (step) in the correct order, instead of being checked and used in the next cycle only, which caused delays, and the delays could affect the steering reduction after each resume.
* To make sure lane change events are only shown after ALC checks, ALC
events are moved to `state_control`.
* Some parts were rearranged/simplified for readability.
* 5-second blinker cooldown was added back for LDW since 2 seconds were
not enough for one-touch lane change when not enabled.
* Functions for ALC and blinker checks were not needed, and were
removed, since now all checks are done inside `state_control`.